+++ Assets/Scripts/LMS/Test.cs
@@ -5,21 +5,20 @@ using UnityEngine;
 public class Test : MonoBehaviour, IDamageable
 {
     LMS.UI.HpBarUI hpBar;
-    int maxHp = 50;
-    int damage = 5;
+    float maxHp = 50;
+    float damage = 5;
     private void Awake()
     {
         hpBar = transform.GetChild(0).transform.GetChild(0).GetComponent<LMS.UI.HpBarUI>();
     }
-    public int GetDamage()
+    public float GetDamage()
     {
         return damage;
     }
 
-    public void TakeDamage(int damage, Vector3 reactVec)
+    public void TakeDamage(float damage, Vector3 reactVec)
     {
         hpBar.UpdateHpBar(damage);
         maxHp -= damage;
     }
-    
 }
